Creating a function that allows users to edit profile info
- Still need to fix bugs
- Testing required

Next step: When email is updated, UI should show the new unverified email and also a badge that tells the user their email is unverified.
If the user hovers the badge, it should give them instructions to go to their old email inbox and accept the new email change.
Once accepted and verified, the banner should disappear.